refactor: extract savings rate chart data
